Move shared network configs to fss-tool
